google-adk-apps/insuraiq/multi_agent_system/insuraiq/requirements.txt
google-adk==1.3.0
google-genai==1.20.0
a2a-sdk==0.2.8
google-cloud-aiplatform[agent-engines]==1.97.0